Moved Window to its own file
